backend/requirements.txt
claude-agent-sdk>=0.1.51
anthropic>=0.49.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0
PyMuPDF>=1.25.0
pydantic>=2.10.0
python-multipart>=0.0.18
aiofiles>=24.1.0
fpdf2>=2.8.0
